+++ inertia/pages/tools/job_calculator/contents/item.ts
@@ -1,0 +1,3 @@
+export function getItemIconUrl(id: string) {
+  return `https://image.palaguidebot.fr/calculator/icons/${id}.webp`
+}